refactor(item.player): refactor item player page
Linked to #33
