+++ public/common/component/device-detection/device-detection.js
@@ -1,3 +1,5 @@
+import angular from 'angular';
+
 class deviceDetectorService {
     constructor($window) {
         this.$window = $window;
@@ -8,5 +10,6 @@ class deviceDetectorService {
     }
 }
 
-angular.module('ps.common.component.device-detection', [])
+export default angular
+    .module('ps.common.component.device-detection', [])
     .service('deviceDetectorService', deviceDetectorService);
