+++ public/item/player/item.player.js
@@ -1,47 +1,13 @@
-class ItemPlayerController {
+import angular from 'angular';
+import AppVideogularConfig from 'config/videogular';
+import AppRouteConfig from 'config/route.config';
+import DeviceDetectionService from 'component/device-detection/device-detection';
+import ItemPlayerController from './item.player.controller';
 
-    constructor(podcast, item, $timeout, deviceDetectorService) {
-        this.item = item;
-        this.item.podcast = podcast;
-        this.$timeout = $timeout;
-
-        this.config = {
-            autoPlay: true,
-            sources: [
-                { src : this.item.proxyURL, type : this.item.mimeType }
-            ],
-            plugins: {
-                controls: {
-                    autoHide: !deviceDetectorService.isTouchedDevice(),
-                    autoHideTime: 2000
-                },
-                poster: this.item.cover.url
-            }
-        };
-    }
-}
-
-angular.module('ps.item.player', [
-    'ngSanitize',
-    'ngRoute',
-    'com.2fdevs.videogular',
-    'com.2fdevs.videogular.plugins.poster',
-    'com.2fdevs.videogular.plugins.controls',
-    'com.2fdevs.videogular.plugins.overlayplay',
-    'com.2fdevs.videogular.plugins.buffering',
-
-    'ps.common.component.device-detection'
+export default angular.module('ps.item.player', [
+    AppRouteConfig.name,
+    AppVideogularConfig.name,
+    DeviceDetectionService.name
 ])
-    .config(($routeProvider) => {
-        $routeProvider.
-            when('/podcasts/:podcastId/item/:itemId/play', {
-                templateUrl: 'item/player/item-player.html',
-                controller: 'ItemPlayerController',
-                controllerAs: 'ipc',
-                resolve : {
-                        item : (itemService, $route) => itemService.findById($route.current.params.podcastId, $route.current.params.itemId),
-                        podcast : (podcastService, $route) => podcastService.findById($route.current.params.podcastId)
-                    }
-                });
-        })
+    .config(ItemPlayerController.routeConfig)
     .controller('ItemPlayerController', ItemPlayerController);
